requirements.txt
Django==5.2.4
PyMySQL
gunicorn
python-dotenv 
asgiref     
django-extensions
openai
elevenlabs